src/requirements.txt
python-telegram-bot==20.0
urllib3==1.26.19
six==1.17.0
boto3==1.24.0
watchtower==2.0.0
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
